Preparation to migrate to SQLAlchemy - Moving the db to sqlalchemy to allow easy migrations and versioning - still built on sqlite
- Introduced a dependency on SQLAlchemy
- Refactored to move all core application components into core.py (db and app)
- Refactored the path wrapper, moving it into utils
- updated flake.nix because apparently venv qt isn't good enough.
