+++ TWMA.py
@@ -1,10 +1,17 @@
 from threading import Thread
 import os as os
+from pathlib import Path
 
-from src.core import app_core, parse_form
+from src.core import app_core, parse_form, app, db
 from src.db import *
 from src.forms import Consent
 from src.view import init_frontend
+
+
+def config_app():
+    db_path = str(Path(os.path.join(os.path.expanduser('~'), 'Documents', 'TWMA_DB')))
+    app.config["SQLALCHEMY_DATABASE_URI"] = f"sqlite:///{db_path}"
+    db.init_app(app)
 
 
 def process_form():
@@ -16,5 +23,6 @@ def process_form():
 if __name__ == '__main__':
     backend = Thread(target=process_form)
     backend.start()
+    config_app()
     init_frontend()  
     backend.join()
