+++ requirements.txt
@@ -4,7 +4,10 @@ bottle==0.13.4
 cffi==2.0.0
 click==8.3.1
 cryptography==46.0.5
+dbus-python==1.4.0
 Flask==3.1.3
+Flask-SQLAlchemy==3.1.1
+greenlet==3.3.2
 itsdangerous==2.2.0
 Jinja2==3.1.6
 MarkupSafe==3.0.3
@@ -14,13 +17,14 @@ pycparser==3.0
 pyinstaller==6.19.0
 pyinstaller-hooks-contrib==2026.2
 pypdf==6.7.5
-PyQt6==6.10.2
+PyQt6==6.9.0
 PyQt6-Qt6==6.10.2
 PyQt6-WebEngine==6.10.0
 PyQt6-WebEngine-Qt6==6.10.2
-PyQt6_sip==13.11.0
+PyQt6_sip==13.10.2
 pywebview==6.1
 QtPy==2.4.3
 setuptools==82.0.0
+SQLAlchemy==2.0.48
 typing_extensions==4.15.0
 Werkzeug==3.1.6
