+++ src/utils.py
@@ -1,3 +1,14 @@
+import sys as sys
+from os import path, getcwd
+from pathlib import Path
 
 def address (street: str, city: str, state: str, zip: str) -> str: 
    return street + ', ' + city + ', ' + state + ' ' + zip if len(street) > 0 else ""
+
+def wrap_path(relative_path: str) -> Path:
+    """if running the app through pyinstaller, this will replace a path of ./something with pyinstaller root/something"""
+    try:
+        base_path = sys._MEIPASS
+    except Exception:
+        base_path = getcwd()
+    return path.join(base_path, relative_path)
